RED_SPIDER/requirements.txt
auto-py-to-exe
bcrypt
subprocess
re
os
time
sys
beautifulsoup4
browser-cookie3
colorama
cryptography
datetime
Counter
sniff
customtkinter
deep-translator
discord
dnspython
exifread
tkinter
shutil
random
ast
string
base64
filedialog
messagebox
hashes
BKDF2HMAC
padding
default_backend
GPUtil
requests
instaloader
smtplib
keyboard
pwinput
MIMEText
opencv-python
phonenumbers
piexif
pillow
psutil
pyautogui
pycryptodome
pyinstaller
pyqt5
pyqtwebengine
pywin32
pyzipper
rarfile
requests
screeninfo
selenium
setuptools
urllib3
whois
whoiam